added dew point; added blank box for null observations

+++ src/main/java/weather_observations/Observations.java
@@ -17,61 +17,61 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
     
     @JsonIgnoreProperties(ignoreUnknown = true)
     public class BarometricPressure {
-      public double value;
+      public Double value;
       public String unitCode;
     }
     
     @JsonIgnoreProperties(ignoreUnknown = true)
     public class HeatIndex {
-      public double value;
+      public Double value;
       public String unitCode;
     }
 
     @JsonIgnoreProperties(ignoreUnknown = true)
     public class SeaLevelPressure {
-      public double value;
+      public Double value;
       public String unitCode;
     }
 
     @JsonIgnoreProperties(ignoreUnknown = true)
     public class Visibility {
-      public double value;
+      public Double value;
       public String unitCode;
     }
 
     @JsonIgnoreProperties(ignoreUnknown = true)
     public class WindSpeed {
-      public double value;
+      public Double value;
       public String unitCode;
 
       /**
        * gives the windspeed in mph instead of kmph
        */
-      public double mph() {
+      public Double mph() {
         return value * 0.621371;
       }
     }
 
     @JsonIgnoreProperties(ignoreUnknown = true)
     public class WindChill {
-      public double value;
+      public Double value;
       public String unitCode;
 
-      public int getTemperature() {
-        if (UnitHandler.getUnit() == TemperatureUnit.Celsius) return (int) value;
-        return (int) (value * 9. / 5.) + 32;
+      public Double getTemperature() {
+        if (UnitHandler.getUnit() == TemperatureUnit.Celsius) return value;
+        return (value * 9. / 5.) + 32;
       }
     }
 
     @JsonIgnoreProperties(ignoreUnknown = true)
     public class WindGust {
-      public double value;
+      public Double value;
       public String unitCode;
     }
 
     @JsonIgnoreProperties(ignoreUnknown = true)
     public class WindDirection {
-      public double value;
+      public Double value;
       public String unitCode;
     }
   }
